Created mortality header file (issue #466)
