+++ ST_globals.h
@@ -30,4 +30,3 @@ extern MortFlagsType  MortFlags;
 extern GlobalType     SuperGlobals;
 
 extern Bool UseGrid;
-extern Bool UseCheatgrassWildfire;
