+++ ST_grid.c
@@ -43,6 +43,7 @@
 #include "ST_initialization.h"
 #include "ST_progressBar.h"
 #include "ST_seedDispersal.h"
+#include "ST_mortality.h"
 
 char sd_Sep;
 
@@ -62,19 +63,17 @@ extern pcg32_random_t grid_rng;         // Gridded mode's unique RNG.
 
 /* We need to seed these RNGs when using the gridded mode but do not use them in this file. */
 extern pcg32_random_t environs_rng;     // Used exclusively in ST_environs.c
-extern pcg32_random_t mortality_rng;    // Used exclusively in ST_mortality.c
 extern pcg32_random_t resgroups_rng;    // Used exclusively in ST_resgroups.c
 extern pcg32_random_t species_rng;      // Used exclusively in ST_species.c
 extern pcg32_random_t markov_rng;       // Used exclusively in SW_Markov.c
 
 extern Bool UseProgressBar;             // From ST_main.c
-extern Bool* _SomeKillage;              // From ST_mortality.c
+extern Bool *_SomeKillage;				// From ST_mortality.c 
 
 /******** Modular External Function Declarations ***********/
 /* -- truly global functions are declared in functions.h --*/
 /***********************************************************/
 //from ST_species.c
-void proportion_Recovery(void);
 void save_annual_species_relsize(void);
 void copy_species(const SpeciesType* src, SpeciesType* dest);
 
@@ -84,14 +83,6 @@ void rgroup_Establish(void);
 void rgroup_IncrAges(void);
 void rgroup_PartResources(void);
 void copy_rgroup(const GroupType* src, GroupType* dest);
-
-//from ST_mortality.c
-void mort_Main(Bool *killed);
-void mort_EndOfYear(void);
-void grazing_EndOfYear(void);
-void _kill_annuals(void);
-void _kill_maxage(void);
-void _kill_extra_growth(void);
 
 //functions from ST_params.c
 void parm_Initialize(void);
@@ -265,10 +256,10 @@ void runGrid(void)
 
 					stat_Collect(year); 		// Update the accumulators
 
-				    _kill_annuals(); 			// Kill annuals
-				    _kill_maxage();             // Kill plants that reach max age
+				    killAnnuals(); 			// Kill annuals
+				    killMaxage();             // Kill plants that reach max age
 					proportion_Recovery(); 		// Recover from any disturbances
-					_kill_extra_growth(); 		// Kill superfluous growth
+					killExtraGrowth(); 		// Kill superfluous growth
 
 				} /* end model run for this cell*/
 			} /* end model run for this row */
